Add a limit flag to label
This commit adds a limit flag to the label command so that we can
control the number of results returned by the search.

This can be useful in situations where we may want to apply
changes in batches or be conservative about the number
of requests we are making against the api.